+++ src/command.rs
@@ -1,24 +1,21 @@
 //! Command parsers and logic.
 
-use anyhow::Result;
 use clap::{CommandFactory, Parser, Subcommand};
-use semver::Version;
-use std::fs::{create_dir_all, File};
-use std::io::{self, BufReader};
-use std::path::{Path, PathBuf};
 
+use std::io;
+use std::path::PathBuf;
+
+mod build_package;
 mod describe;
 mod login;
 mod publish;
 pub mod snowflake;
 mod source;
 mod update;
-use crate::api::{GetPackageVersionResponse, PackageVersion};
-use crate::{api::Client, session};
 
 use self::source::SourceAction;
-use super::codegen::{generate_package, Target};
-use super::descriptor::{DataPackage, Name};
+use super::codegen::Target;
+use super::descriptor::Name;
 use clap_complete::{self, generate, Shell};
 
 #[derive(Subcommand, Debug)]
@@ -57,7 +54,7 @@ enum Command {
     /// usable by the DPM Cloud user that created it; queries made by any other
     /// principal will not be authorized.
     ///
-    /// A package created via (2) is called a standard data package. Queries
+    /// A package created via (2) is called a release data package. Queries
     /// made using it will be authorized if and only if the package's
     /// authorization policy in DPM Cloud allows querying by the given
     /// principal.
@@ -122,27 +119,6 @@ pub struct App {
     command: Command,
 }
 
-/// Reads datapackage.json at path and returns a deserialized instance of DataPackage.
-/// Modified from example code at: https://docs.rs/serde_json/latest/serde_json/fn.from_reader.html#example
-pub fn read_data_package<P: AsRef<Path>>(path: P) -> Result<DataPackage> {
-    let file = File::open(path)?;
-    let reader = BufReader::new(file);
-
-    let data_package = serde_json::from_reader(reader)?;
-    Ok(data_package)
-}
-
-/// Checks that the output directory exists and is accessible.
-fn check_output_dir(p: &Path) {
-    match p.try_exists() {
-        Ok(v) if !v => panic!("Output directory {:?} does not exist", p),
-        Err(e) => {
-            panic!("Error accessing output directory {e}")
-        }
-        _ => {}
-    }
-}
-
 fn print_completions<G: clap_complete::Generator>(gen: G, cmd: &mut clap::Command) {
     generate(gen, cmd, cmd.get_name().to_string(), &mut io::stdout());
 }
@@ -171,40 +147,12 @@ impl App {
                 out_dir,
                 assume_yes,
             } => {
-                // `descriptor` is always defined (possibly via its
-                // default_value), whereas the caller may instead opt to build a
-                // published package via -p. Before reaching this function, clap
-                // will have verified that if -p was given, -d was not given.
-                let build_input: GetPackageVersionResponse = if let Some(package_ref) = package {
-                    let package_identifier: Vec<&str> = package_ref.split('@').collect();
-                    let version: Version = Version::parse(package_identifier[1])
-                        .expect("package identifier `version` is invalid");
-                    let session = session::get_token().expect("unable to get session");
-                    let client = Client::new(&session).expect("unable to get client");
-
-                    client
-                        .get_package_version(package_identifier[0], version)
-                        .await
-                        .expect("failed to fetch package")
-                } else {
-                    let dp = match read_data_package(&descriptor) {
-                        Ok(dp) => dp,
-                        Err(e) => panic!("Error reading {}: {}", descriptor.display(), e),
-                    };
-                    GetPackageVersionResponse {
-                        package_name: dp.name.to_string(),
-                        package_uuid: uuid::Uuid::parse_str(&dp.id.to_string()).unwrap(),
-                        package_description: dp.description.unwrap_or("".into()),
-                        version: PackageVersion {
-                            version: dp.version,
-                            dataset: dp.dataset,
-                        },
-                    }
-                };
-
-                create_dir_all(&out_dir).expect("error creating output directory");
-                check_output_dir(&out_dir);
-                generate_package(&build_input, &target, &out_dir, assume_yes);
+                if let Err(e) =
+                    build_package::build(descriptor, package, target, out_dir, assume_yes).await
+                {
+                    eprintln!("package build failed: {:#}", e);
+                    std::process::exit(1);
+                }
             }
             Command::Login => {
                 if let Err(source) = login::login().await {