+++ src/command/publish.rs
@@ -2,10 +2,10 @@ use std::path::Path;
 
 use anyhow::{Context, Result};
 
-use crate::{api, command::read_data_package, session};
+use crate::{api, descriptor::DataPackage, session};
 
 pub async fn publish(descriptor_path: &Path) -> Result<()> {
-    let package = read_data_package(descriptor_path)
+    let package = DataPackage::read(descriptor_path)
         .with_context(|| format!("Failed to read descriptor at {}", descriptor_path.display()))?;
 
     let token = session::get_token()?;
@@ -17,6 +17,7 @@ pub async fn publish(descriptor_path: &Path) -> Result<()> {
             &package.version,
             &api::CreatePackageVersion {
                 name: &package.name,
+                draft: false,
                 description: &package.description.unwrap_or("".into()),
                 dataset: &package.dataset,
             },